refactor: move tiny_object
