+++ tiny/src/value_object.rs
@@ -1,3 +1,4 @@
 pub mod ast;
 pub mod opcode;
+pub mod tiny_object;
 pub mod token;
