+++ tiny/src/value_object/tiny_object.rs
