+++ tiny/src/vm.rs
@@ -1,5 +1,5 @@
-use crate::tiny_object::TinyObject;
 use crate::value_object::opcode::OpCode;
+use crate::value_object::tiny_object::TinyObject;
 
 #[derive(Debug)]
 pub enum RuntimeError {
